--- navigation_onboarding/main.hpp
@@ -1,3 +1,0 @@
-#include "navigation.hpp"
-
-bool braking_state = false;
