+++ navigation_onboarding/src/main.cpp
@@ -1,0 +1,24 @@
+#include "navigation/navigation.hpp"
+
+bool braking_state = false;
+
+void changeBrakingState(){
+  if(braking_state){braking_state = false;}
+  if(!braking_state){braking_state = true;}
+}
+
+bool doNavigation(){
+  navigation::Navigation nav;
+  bool stop_yet = nav.navigate();
+  return stop_yet;
+}
+
+void brake(){
+  //will send a signal somewhere for us to brake
+}
+
+int main(){
+  //navigation getting called from central logic to estimate our position
+  if(doNavigation()){braking_state = true;}
+  return 0;
+}
