+++ navigation_onboarding/src/navigation/CMakeLists.txt
@@ -1,0 +1,2 @@
+set(target "navigation")
+make_lib(${target} "${CMAKE_CURRENT_SOURCE_DIR}/..")
